Added "Home Wand After Frame" option
When checked, machine will send wand to origin position after the final exposure of each frame. This is useful in cases where the unpredictable final location of the wand might cause shadows that interfere with lighting for subsequent normally lit shots that would be composited later with the light painting shots.